requirements.txt
lxml==4.5.2
PyPDF2==1.26.0
python-docx==0.8.10
